requirements.txt
python-telegram-bot==20.3
logging
datetime
redis
